sirji/vscode-extension/src/py_scripts/requirements-windows.txt
urllib3==2.2.1
pyyaml==6.0.1
sirji-messages==0.0.34
sirji-agents == 0.0.51
